move the rest of channels.go to channels package
